cc65: Modifications to support building entirely using __fastcall__
- Add an option --all-fastcall
- Allow main() to be __fastcall
- Enable cdecl in the compiler (it was mostly there but not enabled)
- Do the right things according to cdecl/fastcall

cc65 is akin to SmallC - it stacks its arguments backwards and passes a size
in y for variadic functions. That means an __fastcall function pointer is not
the same as a function pointer. If you enable all-fastcall then you need to
be careful of this and use cdecl when needed.

Very few apps have function arrays of pointers to variadic functions but
if you do be warned.